cleanup: json_config legacy warnings fixed, style sweep (phase 4)
- json_config.c now warning-free under -Wall -Wextra -Werror:
  packed-member addressing replaced with local-temp parse pattern
  (7 sites in parse_device_config_internal), always-false range
  checks removed (K/W >32767 on uint8, SIM PIN >99999999 on uint16),
  unused octets array dropped
- rf_json.c: explicit (double) casts for float printf args
  (-Wdouble-promotion clean)
- ASCII sweep: Turkish chars removed from rf_dummy.c comments,
  rf_config.h, test file; no tabs, no forbidden calls (heap/errno/
  sprintf) in new modules
- final test run: rf 86/86, rf_json 4/4; full-flag syntax checks pass
- plan doc: phase 4 closed, deferred SCP phase items listed

+++ Application/web-server/rf_json.c
@@ -123,17 +123,17 @@ static unsigned int rfj_emit(char *buf, unsigned int sz, unsigned int pos,
         {
             case RFJ_F1:
                 pos += xsnprintf(&buf[pos], (sz - pos),
-                                 "%s%.1f", (i > 0) ? "," : "", fv);
+                                 "%s%.1f", (i > 0) ? "," : "", (double)fv);
                 break;
 
             case RFJ_F2:
                 pos += xsnprintf(&buf[pos], (sz - pos),
-                                 "%s%.2f", (i > 0) ? "," : "", fv);
+                                 "%s%.2f", (i > 0) ? "," : "", (double)fv);
                 break;
 
             case RFJ_F3:
                 pos += xsnprintf(&buf[pos], (sz - pos),
-                                 "%s%.3f", (i > 0) ? "," : "", fv);
+                                 "%s%.3f", (i > 0) ? "," : "", (double)fv);
                 break;
 
             default:
